Added authorization checks to all note controller endpoints

+++ app/Http/Controllers/Note.php
@@ -8,6 +8,8 @@ class Note extends Controller
 {
     public function show()
     {
+        $this->authorize('viewAny', \App\Models\Note::class);
+
         return view('note', [
             'list' => \App\Models\Note::query()->where('notes.user_id', auth()->id())->paginate(10),
         ]);
@@ -15,11 +17,15 @@ class Note extends Controller
 
     public function create()
     {
+        $this->authorize('create', \App\Models\Note::class);
+
         return view('newnote');
     }
 
     public function newNote(Request $request)
     {
+        $this->authorize('create', \App\Models\Note::class);
+
         $note = \App\Models\Note::query()->create(
             [
                 'name' => $request->input('title'),
@@ -35,6 +41,8 @@ class Note extends Controller
     }
 
     public function save($id, Request $request) {
+        $this->authorize('update', \App\Models\Note::query()->find($id));
+
         $note = \App\Models\Note::query()->where('notes.id', $id)->update(
             [
                 'notes' => $request->input('comment'),
@@ -46,6 +54,8 @@ class Note extends Controller
 
     public function view($id)
     {
+        $this->authorize('view', \App\Models\Note::query()->find($id));
+
         return view('viewnote', [
             'render' => \App\Models\Note::query()->where('id', $id)->find($id)
         ]);
@@ -53,6 +63,8 @@ class Note extends Controller
 
     public function edit($id)
     {
+        $this->authorize('update', \App\Models\Note::query()->find($id));
+
         return view('editnote', [
             'notes' => \App\Models\Note::query()->where('id', $id)->find($id)
         ]);
@@ -60,6 +72,8 @@ class Note extends Controller
 
     public function confirmDelete($id)
     {
+        $this->authorize('delete', \App\Models\Note::query()->find($id));
+
         return view('deletenote', [
             'confirm' => \App\Models\Note::query()->find($id)
         ]);
@@ -67,6 +81,8 @@ class Note extends Controller
 
     public function deleteConfirmed($id)
     {
+        $this->authorize('delete', \App\Models\Note::query()->find($id));
+
         \App\Models\Note::query()->find($id)->delete();
 
         return redirect(route('note.show'));
@@ -74,6 +90,8 @@ class Note extends Controller
 
     public function makePrivate($id)
     {
+        $this->authorize('update', \App\Models\Note::query()->find($id));
+
         \App\Models\Note::findOrFail($id)->update(
             [
                 'private' => true
@@ -85,6 +103,8 @@ class Note extends Controller
 
     public function makePublic($id)
     {
+        $this->authorize('update', \App\Models\Note::query()->find($id));
+
         \App\Models\Note::findOrFail($id)->update(
             [
                 'private' => false
